Replace std io read
