+++ dds/src/rtps_messages/submessage_elements.rs
@@ -1,13 +1,12 @@
 use super::{
     error::{RtpsMessageError, RtpsMessageResult},
-    overall_structure::{Endianness, TryReadFromBytes, Write, WriteIntoBytes},
+    overall_structure::{BufRead, Endianness, TryReadFromBytes, Write, WriteIntoBytes},
     types::FragmentNumber,
     types::ParameterId,
 };
 use crate::transport::types::{Locator, SequenceNumber};
 use alloc::{sync::Arc, vec::Vec};
 use core::ops::Range;
-use std::io::BufRead;
 
 // This files shall only contain the types as listed in the DDS-RTPS Version 2.3
 // 8.3.5 RTPS SubmessageElements
