+++ dds/tests/reliability_tests.rs
@@ -15,7 +15,7 @@ use dust_dds::{
     rtps::types::{PROTOCOLVERSION, VENDOR_ID_S2E},
     rtps_messages::{
         overall_structure::{
-            RtpsMessageHeader, RtpsMessageRead, RtpsMessageWrite, RtpsSubmessageReadKind,
+            BufRead, RtpsMessageHeader, RtpsMessageRead, RtpsMessageWrite, RtpsSubmessageReadKind,
         },
         submessage_elements::{Data, ParameterList, SequenceNumberSet},
         submessages::{ack_nack::AckNackSubmessage, data::DataSubmessage},
@@ -26,7 +26,7 @@ use dust_dds::{
         EntityId, BUILT_IN_READER_WITH_KEY, BUILT_IN_WRITER_WITH_KEY, USER_DEFINED_READER_WITH_KEY,
     },
 };
-use std::io::{BufRead, Read};
+use std::io::Read;
 
 mod utils;
 use crate::utils::domain_id_generator::TEST_DOMAIN_ID_GENERATOR;
